--- a/docroot/modules/custom/va_gov_build_trigger/src/Commands/ContentReleaseCommands.php
+++ b/docroot/modules/custom/va_gov_build_trigger/src/Commands/ContentReleaseCommands.php
@@ -6,10 +6,10 @@ use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\State\StateInterface;
 use Drupal\va_gov_build_trigger\Controller\ContentReleaseNotificationController;
 use Drupal\va_gov_build_trigger\EventSubscriber\ContinuousReleaseSubscriber;
-use Drupal\va_gov_build_trigger\Service\BuildRequesterInterface;
 use Drupal\va_gov_build_trigger\Service\BuildSchedulerInterface;
 use Drupal\va_gov_build_trigger\Service\ReleaseStateManager;
 use Drupal\va_gov_build_trigger\Service\ReleaseStateManagerInterface;
+use Drupal\va_gov_content_release\Request\RequestInterface;
 use Drush\Commands\DrushCommands;
 
 /**
@@ -24,18 +24,18 @@ class ContentReleaseCommands extends DrushCommands {
   protected $releaseStateManager;
 
   /**
-   * The build requester service.
-   *
-   * @var \Drupal\va_gov_build_trigger\Service\BuildRequesterInterface
-   */
-  protected $buildRequester;
-
-  /**
    * The build scheduler service.
    *
    * @var \Drupal\va_gov_build_trigger\Service\BuildSchedulerInterface
    */
   protected $buildScheduler;
+
+  /**
+   * The content release request service.
+   *
+   * @var \Drupal\va_gov_content_release\Request\RequestInterface
+   */
+  protected $requestService;
 
   /**
    * The state management service.
@@ -58,8 +58,8 @@ class ContentReleaseCommands extends DrushCommands {
    *   The release state manager service.
    * @param \Drupal\va_gov_build_trigger\Service\BuildSchedulerInterface $buildScheduler
    *   The build scheduler service.
-   * @param \Drupal\va_gov_build_trigger\Service\BuildRequesterInterface $buildRequester
-   *   The build requester service.
+   * @param \Drupal\va_gov_content_release\Request\RequestInterface $requestService
+   *   The request service.
    * @param \Drupal\Core\State\StateInterface $state
    *   The state service.
    * @param \Drupal\Core\Logger\LoggerChannelFactoryInterface $loggerChannelFactory
@@ -68,13 +68,13 @@ class ContentReleaseCommands extends DrushCommands {
   public function __construct(
     ReleaseStateManagerInterface $releaseStateManager,
     BuildSchedulerInterface $buildScheduler,
-    BuildRequesterInterface $buildRequester,
+    RequestInterface $requestService,
     StateInterface $state,
     LoggerChannelFactoryInterface $loggerChannelFactory
   ) {
     $this->releaseStateManager = $releaseStateManager;
     $this->buildScheduler = $buildScheduler;
-    $this->buildRequester = $buildRequester;
+    $this->requestService = $requestService;
     $this->state = $state;
     $this->logger = $loggerChannelFactory->get('va_gov_build_trigger');
   }
@@ -129,18 +129,6 @@ class ContentReleaseCommands extends DrushCommands {
   }
 
   /**
-   * Request a frontend build (but do not initiate it).
-   *
-   * @command va-gov:content-release:request-frontend-build
-   * @aliases va-gov-content-release-request-frontend-build
-   */
-  public function requestFrontendBuild() {
-    $this->buildRequester->resetFrontendVersion();
-    $this->buildRequester->requestFrontendBuild('Build requested via Drush.');
-    $this->io()->writeln('Frontend build has been requested');
-  }
-
-  /**
    * Make sure builds are going out at least hourly during business hours.
    *
    * @command va-gov:content-release:check-scheduled
@@ -159,7 +147,7 @@ class ContentReleaseCommands extends DrushCommands {
   public function checkStale() {
     if ($this->releaseStateManager->releaseStateIsStale()) {
       $this->resetState();
-      $this->requestFrontendBuild();
+      $this->requestService->submitRequest('Submitting new request due to staleness.');
     }
   }