--- a/docroot/modules/custom/va_gov_build_trigger/src/EventSubscriber/ContinuousReleaseSubscriber.php
+++ b/docroot/modules/custom/va_gov_build_trigger/src/EventSubscriber/ContinuousReleaseSubscriber.php
@@ -6,9 +6,9 @@ use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Datetime\DateFormatterInterface;
 use Drupal\Core\State\StateInterface;
 use Drupal\va_gov_build_trigger\Event\ReleaseStateTransitionEvent;
-use Drupal\va_gov_build_trigger\Service\BuildRequesterInterface;
 use Drupal\va_gov_build_trigger\Service\ReleaseStateManager;
 use Drupal\va_gov_build_trigger\Traits\RunsDuringBusinessHours;
+use Drupal\va_gov_content_release\Request\RequestInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -27,11 +27,11 @@ class ContinuousReleaseSubscriber implements EventSubscriberInterface {
   protected $state;
 
   /**
-   * The build requester service.
+   * The content release request service.
    *
-   * @var \Drupal\va_gov_build_trigger\Service\BuildRequesterInterface
+   * @var \Drupal\va_gov_content_release\Request\RequestInterface
    */
-  protected $buildRequester;
+  protected $requestService;
 
   /**
    * Constructs a new ContentReleaseIntervalSubscriber object.
@@ -40,15 +40,15 @@ class ContinuousReleaseSubscriber implements EventSubscriberInterface {
    *   The state service.
    * @param \Drupal\Component\Datetime\TimeInterface $time
    *   The time service.
-   * @param \Drupal\va_gov_build_trigger\Service\BuildRequesterInterface $buildRequester
+   * @param \Drupal\va_gov_content_release\Request\RequestInterface $requestService
    *   The build requester service.
    * @param \Drupal\Core\Datetime\DateFormatterInterface $dateFormatter
    *   The date formatter service.
    */
-  public function __construct(StateInterface $state, TimeInterface $time, BuildRequesterInterface $buildRequester, DateFormatterInterface $dateFormatter) {
+  public function __construct(StateInterface $state, TimeInterface $time, RequestInterface $requestService, DateFormatterInterface $dateFormatter) {
     $this->state = $state;
     $this->time = $time;
-    $this->buildRequester = $buildRequester;
+    $this->requestService = $requestService;
     $this->dateFormatter = $dateFormatter;
   }
 
@@ -72,7 +72,7 @@ class ContinuousReleaseSubscriber implements EventSubscriberInterface {
 
     if ($is_complete && $is_enabled) {
       $this->runDuringBusinessHours(function () {
-        $this->buildRequester->requestFrontendBuild('Continuous release');
+        $this->requestService->submitRequest('Continuous release');
       });
     }
   }