--- a/docroot/modules/custom/va_gov_build_trigger/src/Plugin/VAGov/Environment/Local.php
+++ b/docroot/modules/custom/va_gov_build_trigger/src/Plugin/VAGov/Environment/Local.php
@@ -3,7 +3,6 @@
 namespace Drupal\va_gov_build_trigger\Plugin\VAGov\Environment;
 
 use Drupal\va_gov_build_trigger\Environment\EnvironmentPluginBase;
-use Drupal\va_gov_build_trigger\Form\LocalBuildTriggerForm;
 
 /**
  * Local Plugin for Environment.
@@ -14,13 +13,6 @@ use Drupal\va_gov_build_trigger\Form\LocalBuildTriggerForm;
  * )
  */
 class Local extends EnvironmentPluginBase {
-
-  /**
-   * {@inheritDoc}
-   */
-  public function getBuildTriggerFormClass() : string {
-    return LocalBuildTriggerForm::class;
-  }
 
   /**
    * {@inheritDoc}