--- a/docroot/modules/custom/va_gov_build_trigger/src/Service/ReleaseStateManager.php
+++ b/docroot/modules/custom/va_gov_build_trigger/src/Service/ReleaseStateManager.php
@@ -326,4 +326,11 @@ class ReleaseStateManager implements ReleaseStateManagerInterface {
     }
   }
 
+  /**
+   * {@inheritDoc}
+   */
+  public function getLastReleaseCompleteTimestamp() : int {
+    return $this->state->get(self::LAST_RELEASE_COMPLETE_KEY, 0);
+  }
+
 }