--- a/docroot/modules/custom/va_gov_content_release/src/Commands/FrontendVersionCommands.php
+++ b/docroot/modules/custom/va_gov_content_release/src/Commands/FrontendVersionCommands.php
@@ -2,6 +2,7 @@
 
 namespace Drupal\va_gov_content_release\Commands;
 
+use Drupal\va_gov_content_release\Frontend\Frontend;
 use Drupal\va_gov_content_release\FrontendVersion\FrontendVersionInterface;
 use Drush\Commands\DrushCommands;
 
@@ -32,33 +33,47 @@ class FrontendVersionCommands extends DrushCommands {
   /**
    * Get the frontend version.
    *
+   * @param string $frontend
+   *   The frontend whose version we are getting.
+   *
    * @command va-gov-content-release:frontend-version:get
    * @aliases va-gov-content-release-frontend-version-get
    */
-  public function get() {
-    $value = $this->frontendVersion->get();
+  public function get(string $frontend = 'content_build') {
+    $frontend = Frontend::fromRawValue($frontend);
+    $value = $this->frontendVersion->getVersion($frontend);
     $this->io()->write($value);
   }
 
   /**
    * Reset the frontend version.
    *
+   * @param string $frontend
+   *   The frontend whose version we are resetting.
+   *
    * @command va-gov-content-release:frontend-version:reset
    * @aliases va-gov-content-release-frontend-version-reset
    */
-  public function reset() {
-    $this->frontendVersion->reset();
+  public function reset(string $frontend = 'content_build') {
+    $frontend = Frontend::fromRawValue($frontend);
+    $this->frontendVersion->resetVersion($frontend);
     $this->io()->success('Frontend version reset.');
   }
 
   /**
    * Set the frontend version.
    *
+   * @param string $frontend
+   *   The frontend whose version we are getting.
+   * @param string $version
+   *   The version to set.
+   *
    * @command va-gov-content-release:frontend-version:set
    * @aliases va-gov-content-release-frontend-version-set
    */
-  public function set($version) {
-    $this->frontendVersion->set($version);
+  public function setVersion(string $frontend = 'content_build', $version = '_default') {
+    $frontend = Frontend::fromRawValue($frontend);
+    $this->frontendVersion->setVersion($frontend, $version);
     $this->io()->success('Frontend version set to ' . $version);
   }