Closes #6434.

## QA Steps

I did the following:

### A normal build

![Screenshot 2023-11-08 at 3 05 13 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/1318579/d9c040cf-71b8-431b-b68d-d111c5811139)

![Screenshot 2023-11-08 at 3 05 23 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/1318579/c3d00f57-8ad2-4fe1-9e97-d34e0f3e9e2d)


```
Wed Nov  8 06:26:03 PM UTC 2023
content-build version: __default
vets-website version: __default
==> Starting a frontend build. This file will be updated as the build progresses.
==> Resetting VA repos to default versions
Gathering patches for root package.
Installing dependencies from lock file (including require-dev)
Verifying lock file contents can be installed on current platform.
Package operations: 1 install, 0 updates, 0 removals
  - Syncing va-gov/content-build (v0.0.3383) into cache
Gathering patches for root package.
Gathering patches for dependencies. This might take a minute.
  - Installing va-gov/content-build (v0.0.3383): Cloning ec4aa8360e from cache
Package doctrine/reflection is abandoned, you should avoid using it. Use roave/better-reflection instead.
Package php-http/message-factory is abandoned, you should avoid using it. Use psr/http-factory instead.
Package simplesamlphp/simplesamlphp-module-authfacebook is abandoned, you should avoid using it. No replacement was suggested.
Package simplesamlphp/simplesamlphp-module-authwindowslive is abandoned, you should avoid using it. No replacement was suggested.
Package simplesamlphp/simplesamlphp-module-oauth is abandoned, you should avoid using it. No replacement was suggested.
Package simplesamlphp/simplesamlphp-module-riak is abandoned, you should avoid using it. No replacement was suggested.
Package simplesamlphp/simplesamlphp-module-sanitycheck is abandoned, you should avoid using it. No replacement was suggested.
Package simplesamlphp/twig-configurable-i18n is abandoned, you should avoid using it. No replacement was suggested.
Package symfony/debug is abandoned, you should avoid using it. Use symfony/error-handler instead.
Package twig/extensions is abandoned, you should avoid using it. No replacement was suggested.
Package webmozart/path-util is abandoned, you should avoid using it. Use symfony/filesystem instead.
Generating optimized autoload files
113 packages you are using are looking for funding.
Use the `composer fund` command to find out more!
==> Using default content-build version
==> Installing yarn dependencies
> # Prepare content-build project.
> ! ./scripts/should-run-directly.sh || ./scripts/web-install.sh
/var/lib/tugboat/bin /var/lib/tugboat /var/lib/tugboat
/var/lib/tugboat /var/lib/tugboat
Found '/var/lib/tugboat/.nvmrc' with version <16.19.1>
v16.19.1 is already installed.
Now using node v16.19.1 (npm v8.19.3)
Node v16.19.1
NPM 8.19.3
Yarn 1.19.1
Installed yarn version '1.19.1' matches the required version '1.19.1'.
/var/lib/tugboat/web /var/lib/tugboat /var/lib/tugboat
Found '/var/lib/tugboat/web/.nvmrc' with version <14.15.0>
v14.15.0 is already installed.
Now using node v14.15.0 (npm v6.14.8)
yarn run v1.19.1
warning ../../../../package.json: License should be a valid SPDX license expression
$ ./script/install-repos.sh
Cloning into '../vagov-content'...
Cloning into '../vets-website'...
Cloning into '../vets-api'...
Updating files: 100% (8676/8676), done.
Done in 20.26s.
Warning: Ignoring extra certs from `/etc/pki/tls/certs/ca-bundle.crt`, load failed: error:02001002:system library:fopen:No such file or directory
yarn install v1.19.1
warning ../../../../package.json: License should be a valid SPDX license expression
[1/4] Resolving packages...
warning Resolution field "lodash@4.17.21" is incompatible with requested version "lodash@^2.4.1"
[2/4] Fetching packages...
info fsevents@2.3.2: The platform "linux" is incompatible with this module.
info "fsevents@2.3.2" is an optional dependency and failed compatibility check. Excluding it from installation.
[3/4] Linking dependencies...
warning " > @department-of-veterans-affairs/eslint-plugin@1.1.0" has incorrect peer dependency "eslint-plugin-jsx-a11y@^6.5.1".
warning " > @department-of-veterans-affairs/eslint-plugin@1.1.0" has incorrect peer dependency "eslint-plugin-sonarjs@^0.12.0".
warning " > @department-of-veterans-affairs/eslint-plugin@1.1.0" has incorrect peer dependency "eslint-plugin-unicorn@^35.0.0".
warning " > @department-of-veterans-affairs/web-components@4.9.6" has unmet peer dependency "@department-of-veterans-affairs/formation@^7.0.1".
warning " > @department-of-veterans-affairs/web-components@4.9.6" has unmet peer dependency "i18next@*".
warning " > @department-of-veterans-affairs/web-components@4.9.6" has unmet peer dependency "i18next-browser-languagedetector@*".
warning " > eslint-plugin-prettier@4.0.0" has incorrect peer dependency "prettier@>=2.0.0".
warning " > mocha-snapshots@4.2.0" has incorrect peer dependency "mocha@^4.0.1".
[4/4] Building fresh packages...
$ husky install
Warning: Ignoring extra certs from `/etc/pki/tls/certs/ca-bundle.crt`, load failed: error:02001002:system library:fopen:No such file or directory
husky - Git hooks installed
Done in 12.46s.
/var/lib/tugboat /var/lib/tugboat
> ./scripts/should-run-directly.sh || ddev composer va:web:install --
==> Using default vets-website version
...
Done in 2644.99s.
/var/lib/tugboat /var/lib/tugboat
Replacing s3 address with local in generated files.
> ./scripts/should-run-directly.sh || ddev composer va:web:build --
==> Build complete
```

### A build with specified `content-build` and `vets-website` versions:

![Screenshot 2023-11-08 at 2 20 22 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/1318579/24c15371-13b4-4ce2-bce6-77066d544205)

![Screenshot 2023-11-08 at 2 20 31 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/1318579/1f5216ab-e1fe-4b8b-bc09-73829333eb67)

```
Wed Nov  8 07:20:33 PM UTC 2023
content-build version: 1783
vets-website version: 26628
...
==> Checking out the requested frontend version
From https://github.com/department-of-veterans-affairs/content-build
 * branch                refs/pull/1783/head -> FETCH_HEAD
Previous HEAD position was ec4aa8360 fix naked link, remove aria label, use va-link component (#1782)
HEAD is now at 314282203 Enable form 21-0966
...
==> Checking out the requested vets-website version
From https://github.com/department-of-veterans-affairs/vets-website
 * branch                  refs/pull/26628/head -> FETCH_HEAD
Note: switching to 'FETCH_HEAD'.

You are in 'detached HEAD' state. You can look around, make experimental
changes and commit them, and you can discard any commits you make in this
state without impacting any branches by switching back to a branch.

If you want to create a new branch to retain commits you create, you may
do so (now or later) by using -c with the switch command. Example:

  git switch -c <new-branch-name>

Or undo this operation with:

  git switch -

Turn off this advice by setting config variable advice.detachedHead to false

HEAD is now at bbe24f57fd Reenable deep linking PACT Act test after devops change
...
Done in 2324.76s.
/var/lib/tugboat /var/lib/tugboat
Replacing s3 address with local in generated files.
> ./scripts/should-run-directly.sh || ddev composer va:web:build --
==> Build complete
```

## Automated Tests

See this video: https://pr15523-fd72c9bfvnmyedcrkvhxbnlciqgevv3m.ci.cms.va.gov/cypress/videos/platform/content_release.feature.mp4