@ndouglas 2023-11-08
@jilladams No need to test anything, but I'd appreciate it if you could tell me whether this roughly matched the experience you had in mind for this functionality. If there are any small changes you'd like 😓, I could roll them into a followup ticket.