@ndouglas 2023-11-08
Not seeing any accessibility issues.

These two are for best practices and related to the toolbar, not my changes.

![Screenshot 2023-11-08 at 3 18 01 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/1318579/06e9ebaf-4b3a-44be-b71b-f156c123f438)