@jilladams 2023-11-08
The interface makes sense, conceptually. I think the help text is confusing. I don't know if this is within our control (and it may not be worth the effort to change), but: for example, if I type `clp` in the vets-website picker:  I can select a 3 year old PR, with a deleted branch: 

![Screenshot 2023-11-08 at 2 20 26 PM](https://github.com/department-of-veterans-affairs/va.gov-cms/assets/85581471/8e5bd3ab-b2b2-45b1-87c6-d8646af72ac1)

PR: https://github.com/department-of-veterans-affairs/vets-website/pull/16896
Branch: https://github.com/department-of-veterans-affairs/vets-website/tree/kelsonic-23418-campaign-mission-act-analytics

Would the build succeed if I did that? What branch would it use? ( I didn't do a content release to find out, because that takes a long time.)

Again: that may already be default behavior for the content-build picker, and I think by and large people bothering to use this field will specify a branch they care about that still exists.