@jilladams 2023-11-08
Nevermind - I checked the prod mirror tugboat and this is how it already is for the content-build one so: feel free to call it good enough! Woo!